Fix the over-calling of VoiceSearch Logging
bug: 3035228 This change will reduce the number of times VoiceSearch app
is called to log IME events.  This is because it 1) only
flushes the logs when there is relevant logging information
available and 2) sets a flag in UserHappinessSignals so
that other apps (eg: gmail/sms) can know if there is
relevant logging information before sending the logging
intent to voice search.

Change-Id: I708fea5a38f854091f1101e4f6a3db46296f3e92