Add a first version of Hebrew and Arabic keyboard.
The Arabic keyboard is more or less based on PC keyboard and other
sources. It has no means to input diacritics at the moment.
Some data for the Hebrew keyboard was already there, but not used.
This change splits phone and tablet layout and consolidates the
tablet one to something that seems sensible. It can't input diacritics
either at the moment.

Other shortcomings affect those keyboards. For example, normal
Hebrew keyboards switch to capital QWERTY keyboard with shift on,
as there are no capital Hebrew characters, but this version does not
feature this.

It should be mostly possible already to enter Arabic or Hebrew with
these keyboards however.

Change-Id: Ic4df2a9a77ffd03c4f9ee2c47e03c0f43f8e48ae