Fix a bug where dead space would appear on the keyboard.
This change fixes a bug where there would be pixels on the keyboard
that would not return a key although it's between two keys - or even
right on top of one.
This change makes it so that the closest key to the touch - within a
certain threshold - is always returned, regardless of whether the
touch is inside or not.

Bug: 4348994

Change-Id: I8f6102d6787eb025cc3c50a26d3a475aeafc4b64