Move the settings test to a more appropriate place.
This change refactors the dictionary selection code so that the
cached dictionary files list and the settings tests are more
cleanly separated.
This will also help with future refactorings that will test for
the presence of the main dictionary and insert the fall back if
it's not supplied by the dictionary pack.

Bug: 5095140
Change-Id: I8d7caad7c054031df71fe78b043801a774d50f65