Filter out of spellchecking words that don't look like words
Note that this affects only the results of the spell checker if
actually passed such a word. For example, the spell checker will
not flag "http://oju" as a typo, because it looks like a URL.
But in the current implementation, TextView passes "http" and
"oju" separately, so "oju" is still flagged as a typo.

Bug: 5281875
Change-Id: I9d721fd3af34edc51b11908cf2e8fe994b164242