Don't reset animation alpha value when keyboard is set
The keys' blinking that mentioned in the bug is caused by switching
keyboard layout resets the animation's alpha value to opaque but the
animation is still running and changing the alpha value
asynchronously.

I think that switching keyboard layout between alphabet and symbols
doesn't imply that the user stops typing. So the keyboard view should
continue typing state timer to keep animations' alpha values changing
and never reset the value.

Bug: 6174273
Change-Id: Id795feaf44750358f30c1b3dc8e783a7e62aefe8