Move the UserUnigramDictionary functionality over
UserBigramDictionary now assumes both functionalities. It will
be renamed to UserHistoryDictionary in a future change.

There are several reasons to do this. First, there is a lot of
duplicate code in User{Unigram,Bigram}Dictionaries that are
factored by the few lines of code in this change. Also, other
dictionaries like BinaryDictionary or ContactsDictionary all
assume both responsibilities, as should be the case
theoretically.

It is also possible to do this because previous versions don't
write any unigram data that we'd want to reuse. For even older
versions that do write data, we can't really make any sense out
of it. Bigram data however can be useful, and this allows us to
reuse it easily.

Change-Id: I755525f92744e1536eaef097527e8151b7859a30