Correct a shaky processing and move it to a better place
The old code would remove caps from the first letter if
auto-capsed but that makes very little sense when the word
is camel-cased. Also, it would not correctly handle
surrogate pairs, and would require a direct reference to
LatinIME and a specific method to do the processing in an
unexpected place.

Change-Id: I416d6a805242788a2473f007ca7452c9fe3f5205