Consolidate auto-correction settings.
For some reason, there were several ways, not exactly identical,
to get this setting. The ones that used mAutoCorrectionEnabled
would kick in when the input field was specifying no correction,
so it would be a little strange (although harmless in the practice
because the settings set in this way would not get used later,
because the correct test would be done at that time).

Also perform a very small refactoring

Change-Id: Ica9f32b238d98009ae1852d3c1e940398f5d341c