Cleanup unused code
The Italian dictionary has been amended to contain words that
include apostrophes, and accordingly the apostrophe is now
excluded from word separators in Italian.
This code was only used for Italian, and in the general case
it does the wrong thing; it was only needed by limitations of
the old dictionary.

Bug: 6929108
Change-Id: I7ff4615e01899a1cebed2c4c51679cd0746740fe