Fix a logic flaw.
This test is seriously wrong : although we do have a method that
does exactly the right thing, this looks at a very specific, unrelated
variable state in hope it happens to be set only in the relevant
cases by the play of logic.
I don't think this changes anything concretely, as the assumption
was probably true in the practice ; still, this code is very hard to
read and prone to unexpected breakage.

Change-Id: I77ab576f90762429fcc752df31ed30e1c81871dc