Fix crashes related to setSpan(-1 ... -1) errors
Fixes b/7652661

In some circumstnaces, when hiding the window, the logger tries to capture the TextView contents
and restore the cursor position based on the values stored in LatinIME.  However, the IME may not
know the cursor position.

This fix does two things 1) doesn't try to reposition the cursor if the TextView contents are not
to be captured, and 2) if the cursor position is unknown (i.e., == -1), doesn't try to move it back.

Change-Id: I3976aaccf58c18f892c3a8530218ae52bcac9989