[Rlog60] Log if a logUnit is uncommitted
Normally logUnits are committed to a log when the word that their data
corresponds to is complete.   However, if the user reverts a word, or goes
back to edit it, then the system may "uncommit" the logUnit, and append
additional editing steps to the end of it.  When this happens, we want to
make a note of it in the log, which is what this patch does.

Change-Id: If2bbb948469824c76facf5f06ed8a6da8ff9777c